+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#202](https://github.com/os2display/display-api-service/pull/202)
+  - Fixed ScreenUser blamable identifier.
+
 ## [2.0.1] - 2024-04-09
 
 - [#201](https://github.com/os2display/display-api-service/pull/201)